refactor: refactor package
