+++ market/build.gradle
@@ -37,11 +37,6 @@ dependencies {
 	implementation 'org.postgresql:postgresql:42.5.1'
 
 	testImplementation 'org.springframework.boot:spring-boot-starter-test:2.7.6'
-//	testImplementation 'org.junit.jupiter:junit-jupiter-params:5.9.1'
-//	testImplementation 'org.junit.jupiter:junit-jupiter-engine:5.9.1'
-//	testImplementation 'org.assertj:assertj-core:3.23.1'
-//	testImplementation 'org.mockito:mockito-core:4.9.0'
-//	testImplementation 'org.mockito:mockito-junit-jupiter:4.9.0'
 }
 
 tasks.named('test') {
