--- market/src/main/java/by/bobrovich/market/cache/data/request/RequestAlcoholDto.java
@@ -1,15 +1,0 @@
-package by.bobrovich.market.cache.data.request;
-
-import lombok.Builder;
-
-import java.math.BigDecimal;
-
-@Builder
-public record RequestAlcoholDto(
-        String name,
-        String country,
-        double vol,
-        BigDecimal price,
-        int quantity
-) {
-}
