+++ market/src/main/java/by/bobrovich/market/dao/InMemoryAlcoholDao.java
@@ -1,8 +1,8 @@
-package by.bobrovich.market.cache.dao;
+package by.bobrovich.market.dao;
 
 import by.bobrovich.market.cache.annotation.Cache;
-import by.bobrovich.market.cache.dao.api.AlcoholDao;
-import by.bobrovich.market.cache.entity.Alcohol;
+import by.bobrovich.market.dao.api.AlcoholDao;
+import by.bobrovich.market.entity.Alcohol;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Component;
