+++ market/src/main/java/by/bobrovich/market/dao/postgresql/JdbcAlcoholDao.java
@@ -1,10 +1,10 @@
-package by.bobrovich.market.cache.dao;
+package by.bobrovich.market.dao.postgresql;
 
 import by.bobrovich.market.cache.annotation.Cache;
-import by.bobrovich.market.cache.dao.api.AlcoholDao;
-import by.bobrovich.market.cache.entity.Alcohol;
-import by.bobrovich.market.cache.exception.AlcoholNotFoundException;
-import by.bobrovich.market.cache.exception.AlcoholSqlException;
+import by.bobrovich.market.dao.api.AlcoholDao;
+import by.bobrovich.market.entity.Alcohol;
+import by.bobrovich.market.exceptions.AlcoholNotFoundException;
+import by.bobrovich.market.exceptions.AlcoholSqlException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.stereotype.Repository;
@@ -17,7 +17,7 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 
-import static by.bobrovich.market.cache.dao.util.AlcoholQueryUtil.*;
+import static by.bobrovich.market.dao.postgresql.util.AlcoholQueryUtil.*;
 
 @RequiredArgsConstructor
 @Repository
